* persist serial number as decimal in DB for correct Ordering * add reception time to response data

+++ signing/src/main/java/dev/mieser/tsa/signing/serial/SerialNumberGenerator.java
@@ -1,7 +1,5 @@
 package dev.mieser.tsa.signing.serial;
 
-import java.math.BigInteger;
-
 /**
  * Generates unique serial numbers for TSP responses.
  */
@@ -10,8 +8,7 @@ public interface SerialNumberGenerator {
 
     /**
      * @return A unique serial number.
-     * @implSpec The returned serial number cannot be {@code null}.
      */
-    BigInteger generateSerialNumber();
+    long generateSerialNumber();
 
 }
